# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.BuildingResponse:
    properties:
      address:
        type: string
      branch_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      organization_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.CreateBuildingRequest:
    properties:
      address:
        type: string
      branch_id:
        type: string
      description:
        type: string
      name:
        type: string
      organization_id:
        type: integer
    required:
    - address
    - branch_id
    - name
    - organization_id
    type: object
...
  dto.ListBuildingsResponse:
    properties:
      buildings:
        items:
          $ref: '#/definitions/dto.BuildingResponse'
        type: array
      total:
        type: integer
    type: object
...
  dto.UpdateBuildingRequest:
    properties:
      address:
        type: string
      branch_id:
        type: string
      description:
        type: string
      name:
        type: string
      organization_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a template API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Template API
  version: "1.0"
paths:
  /buildings:
    get:
      description: Retrieve a list of buildings with pagination and sorting support.
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: 'Sort fields (format: field1:asc,field2:desc) - Available fields:
          first_name, last_name, email, created_at'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of buildings with pagination information
          schema:
            $ref: '#/definitions/dto.ListBuildingsResponse'
        "400":
          description: Invalid input parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List buildings
      tags:
      - building
    post:
      consumes:
      - application/json
      description: Create a new building with the provided details.
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: Building information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBuildingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created building
          schema:
            $ref: '#/definitions/dto.BuildingResponse'
        "400":
          description: Invalid input, bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create building
      tags:
      - building
  /buildings/{id}:
    delete:
      description: Delete a building by its ID
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: Building ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Building not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete building
      tags:
      - building
    get:
      description: Retrieve a building's details using its ID
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: building ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Building details
          schema:
            $ref: '#/definitions/dto.BuildingResponse'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Building not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get building by ID
      tags:
      - building
    put:
      consumes:
      - application/json
      description: Update a building's details by its ID
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: Building ID
        in: path
        name: id
        required: true
        type: string
      - description: Building information to be updated
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBuildingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated Building details
          schema:
            $ref: '#/definitions/dto.BuildingResponse'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Building not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update building
      tags:
      - building
  /buildings/search:
    get:
      description: Search for a building by name or email with optional case sensitivity,
        pagination, and sorting.
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: Search query (e.g., building name )
        in: query
        name: search
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Case sensitive search (default: false)'
        in: query
        name: case_sensitive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of buildings matching search criteria with pagination
            info
          schema:
            $ref: '#/definitions/dto.ListBuildingsResponse'
        "400":
          description: Invalid input parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search building
      tags:
      - building
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Enter your bearer token in the format **Bearer &lt;token&gt;**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"